09-ai-db-query-optimizer/requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sqlparse>=0.5.0
pandas>=2.0.0
